+++ backend/greenwebanalyzer/report.py
@@ -1,6 +1,7 @@
 # Selenium Wire
 from seleniumwire import webdriver
 from selenium.webdriver import ChromeOptions
+from selenium.webdriver.support.wait import WebDriverWait
 from seleniumwire.utils import decode
 
 # Time
@@ -12,7 +13,7 @@ import pytz
 import os
 
 # Criterias
-from .criteria import criteria_requests, criteria_img_types, criteria_img_compression, criteria_redirects, criteria_minified_files
+from .criteria import criteria_requests, criteria_img_types, criteria_img_compression, criteria_redirects, criteria_minified_files, criteria_img_lazy_loaded
 
 from .utils import create_folder, delete_folder
 
@@ -46,14 +47,16 @@ class Report:
         self.folder_name = f"request-{time.time()}"
 
         self.driver = webdriver.Chrome(options=self.options)
-        self.driver.set_window_size(1920, 1080)
 
         self.app = app
 
     def request_page(self) -> None:
         del self.driver.requests
         self.driver.start_session(self.options.to_capabilities())
+        self.driver.set_window_size(1920, 1080)
+        self.driver.implicitly_wait(10)
         self.driver.get(self.url)
+        time.sleep(3)
 
     def save_page(self) -> None:
         self.requests = list()
@@ -194,8 +197,11 @@ class Report:
         # Criteria 3: Image Compression
         criteria_3 = criteria_img_compression(self.file_paths['img'])
 
+        # Criteria 3: Image Lazy loaded
+        criteria_4 = criteria_img_lazy_loaded(self.driver, self.requests)
+
         # Criteria 4: Minify files
-        criteria_4 = criteria_minified_files(self.file_paths)
+        criteria_5 = criteria_minified_files(self.file_paths)
 
         # Combine
         criterias = [
@@ -203,7 +209,8 @@ class Report:
             criteria_1,
             criteria_2,
             criteria_3,
-            criteria_4
+            criteria_4,
+            criteria_5,
         ]
 
         # End
@@ -217,7 +224,7 @@ class Report:
         metrics = {
             "size": self.full_size,
             "requests": amount_requests,
-            "potential_savings": criteria_2['details']['total_savings'] + criteria_4['details']['total_savings']
+            "potential_savings": criteria_2['details']['total_savings'] + criteria_5['details']['total_savings']
         }
 
         # Create report